pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp_eval_llm"
version = "1.0.0"
description = "Model Context Protocol (MCP) based LLM evaluation pipeline"
readme = "README.md"
requires-python = ">=3.12"
license = {file = "LICENSE"}
authors = [
    {name = "Zhiwei Liu", email = "zhiweiliu@salesforce.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # Core evaluation framework
    "anthropic>=0.8.0",
    "openai>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
    "jsonschema>=4.17.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.6.0",
    "aiohttp>=3.11.16",
    "serpapi>=0.1.5",
    "yfinance>=0.2.61",
    # Backend API dependencies
    "flask>=2.3.3",
    "flask-cors>=4.0.0",
    "werkzeug>=2.3.7",
    "psutil>=5.9.8",
    "click>=8.1.7",
    "colorlog>=6.7.0",
    "cerberus>=1.3.4",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    # Core development tools
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    
    # Backend development tools
    "pytest-flask>=1.3.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
mcp-eval = "mcp_eval_llm.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["mcp_eval_llm*"]

[tool.setuptools.package-data]
mcp_eval_llm = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 
